requirements.txt
langchain>=0.1.0
langchain-google-genai>=1.0.0
langgraph>=0.0.40
python-dotenv>=1.0.0
langchain-community>=0.1.0
faiss-cpu>=1.7.4
langchain-text-splitters>=0.1.0

